# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WPMgr API
  version: 0.0.0
  description: |
    Control-plane REST API for WPMgr. This is the single source of truth for
    both the Go server types and the generated TypeScript client. Real paths
    and schemas are filled in during Phase 4 (skeleton) and Phase 5 (features).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/sites/{siteId}:
    get:
      operationId: getSite
      summary: Get a site by ID
      tags: [sites]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: The site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSite
      summary: Delete a site
      tags: [sites]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "204":
          description: Site deleted
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/sites/{siteId}/diagnostics:
    get:
      operationId: getSiteDiagnostics
      summary: Latest site diagnostics by category
      description: |
        Returns one card per category (14 total) carrying the latest payload
        the agent shipped + the collection/freshness timestamps. Categories the
        agent has never reported for are returned with a null payload so the UI
        can render an "awaiting first sync" placeholder.
      tags: [diagnostics]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: Site diagnostics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDiagnosticsList"
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/sites/{siteId}/diagnostics/refresh:
    post:
      operationId: refreshSiteDiagnostics
      summary: Enqueue an on-demand diagnostics command to the agent
      description: |
        Enqueues a signed `diagnostics` command to the agent. The agent runs
        the 14-category collector immediately and pushes the result back to
        POST /agent/v1/diagnostics. Returns 202 on accept. Returns 503 with
        code `diagnostics_refresh_unwired` when the CP->agent commander has
        not yet been wired (V1).
      tags: [diagnostics]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "202":
          description: Refresh enqueued
        "503":
          description: Refresh is not yet wired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/sites/{siteId}/errors:
    get:
      operationId: listSitePHPErrors
      summary: List fingerprint-grouped PHP errors for a site
      description: |
        Returns the agent-captured PHP errors for the site, grouped by md5
        fingerprint. Each row carries the occurrence count, first/last seen
        timestamps, and the silenced flag.
      tags: [diagnostics]
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Only rows whose last_seen_at is after this timestamp.
        - name: silenced
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
          description: Filter by silenced state. Omit for both.
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: PHP error list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PHPErrorList"
  /api/v1/sites/{siteId}/errors/{md5}/silence:
    post:
      operationId: silenceSitePHPError
      summary: Silence (or unsilence) a fingerprint-grouped PHP error
      description: |
        Toggles the silenced flag on a (site, md5) error row. The agent
        continues to count silently on its side; the CP UI hides silenced
        rows by default.
      tags: [diagnostics]
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - name: md5
          in: path
          required: true
          schema:
            type: string
          description: The md5 fingerprint.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PHPErrorSilence"
      responses:
        "204":
          description: Silence flag updated
        "404":
          description: Error not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  # ── ADR-037 Sprint 3 — WordPress activity log (tamper-evident) ──────────────
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
...
    SiteId:
      name: siteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    AgentSignature:
      type: apiKey
      in: header
      name: X-WPMgr-Signature
      description: |
        Ed25519 signed-request authentication for agent->control-plane calls. The
        agent sends four headers: X-WPMgr-Agent-Key (its base64 Ed25519 public
        key, identifying the enrolled site), X-WPMgr-Timestamp (Unix seconds),
        X-WPMgr-Nonce (a unique per-request token), and X-WPMgr-Signature (the
        base64 Ed25519 signature over the canonical message
        "METHOD\nPATH\nTIMESTAMP\nNONCE\nhex(sha256(body))"). The control plane
        rejects stale timestamps and replayed nonces. (apiKey type is the closest
        OpenAPI 3.1 primitive for this header-carried signature scheme.)
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          type: object
          additionalProperties: true
...
    Site:
      type: object
      required:
        [id, tenant_id, url, name, status, wp_version, php_version, health_status, multisite, tags, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        name:
          type: string
        status:
          type: string
          enum: [pending, active, error, disabled]
        wp_version:
          type: string
        php_version:
          type: string
        health_status:
          type: string
          enum: [unknown, healthy, unreachable]
        server_info:
          type: string
        multisite:
          type: boolean
        active_theme:
          type: string
        tags:
          type: array
          items:
            type: string
        enrolled:
          type: boolean
          description: Whether an agent has enrolled this site.
        enrolled_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        components:
          $ref: "#/components/schemas/SiteComponents"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SiteComponents:
      type: object
      description: Installed plugin/theme inventory pushed by the agent.
      properties:
        plugins:
          type: array
          items:
            $ref: "#/components/schemas/SiteComponent"
        themes:
          type: array
          items:
            $ref: "#/components/schemas/SiteComponent"
        core_update:
          type: object
          nullable: true
          description: When set, WordPress core has an update available.
          required: [new_version, current_version]
          properties:
            new_version:
              type: string
            current_version:
              type: string
    SiteComponent:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
        name:
          type: string
        version:
          type: string
        active:
          type: boolean
        available_update:
          type: object
          nullable: true
          description: When set, an update is available for this plugin/theme.
          required: [new_version]
          properties:
            new_version:
              type: string
            package:
              type: string
              nullable: true
            tested:
              type: string
              nullable: true
            requires_php:
              type: string
              nullable: true
...
    SiteDiagnosticsCard:
      type: object
      required: [category, payload, fresh]
      properties:
        category:
          type: string
          description: |
            One of: identity, php, mysql, filesystem, http, cron, themes,
            plugins, users, security, https, mail, performance, hosting,
            wp_native. The first 14 are the WPMgr-extra leapfrog collector
            (the 9-card legacy grid + the ribbon-summary entries); the 15th
            `wp_native` carries the verbatim WP_Debug_Data::debug_data()
            dump (full Site Health > Info parity) and is rendered as four
            additional cards: Directory Sizes, Media Handling, Filesystem
            Permissions, WordPress Constants.
        payload:
          description: |
            The raw JSON sub-blob the agent shipped for this category, or
            null when the agent has never shipped a payload for it yet.
            Shape is category-specific and deliberately tolerant — the UI
            parses it lazily so the agent can evolve the shape without a
            wire schema change.
        collected_at:
          type: string
          format: date-time
          description: Agent-side collection timestamp (UTC).
        received_at:
          type: string
          format: date-time
          description: CP-side ingestion timestamp (UTC).
        fresh:
          type: boolean
          description: |
            True when the agent's `collected_at` is within the freshness
            tolerance (24h + 2h slack). Card UIs render a stale warning
            badge when false.
    SiteDiagnosticsList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SiteDiagnosticsCard"
    PHPError:
      type: object
      required:
        - id
        - md5
        - code
        - severity
        - message
        - file
        - line
        - request_path
        - first_seen_at
        - last_seen_at
        - occurrence_count
        - silenced
      properties:
        id:
          type: string
          description: The CP-side row UUID (stringified).
        md5:
          type: string
          description: The agent-side dedup fingerprint (md5(code:file:line:message)).
        code:
          type: integer
          description: PHP error code (E_* constant value).
        severity:
          type: string
          description: One of fatal, warning, notice, deprecated, bootstrap, unknown.
        message:
          type: string
        file:
          type: string
        line:
          type: integer
        request_path:
          type: string
        first_seen_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        occurrence_count:
          type: integer
          format: int64
        silenced:
          type: boolean
    PHPErrorList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PHPError"
    PHPErrorSilence:
      type: object
      properties:
        silenced:
          type: boolean
          default: true
          description: Whether to silence (default true).
    # ── ADR-037 Sprint 3 — WordPress activity log ──────────────────────────
...